requirement.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
gunicorn==23.0.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.21.0
packaging==24.1
pillow==10.3.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sympy==1.13.3
tokenizers==0.20.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
